\name{AMC_Amb}
\docType{data}
\alias{AMC_Amb}
\title{Failures times of an AMC Ambassador car}
\description{
	18 failures times of AMC Ambassador car owned by the Ohio state government.
}
\usage{AMC_Amb}
\format{A vector of successive failure times.}
\references{
  \itemize{
  	\item C.W. Ahn, K.C. Chae and G.M. Clark (1998). Estimating parameters of the power-law process with two measures of time. \emph{Journal of Quality Technology} 30(2), pp. 127--132.
  	\item M. Guida and G. Pulcini (2006). Bayesian analysis of repairable systems showing a bounded failure intensity. \emph{Reliability Enginering and System Safety} 91, pp. 828--838.
  	\item F. Corest, L. Doyen and O. Guadoin (2012). Bayesian analysis of ARA imperfect repair models. \emph{Communications in Statistics - Theory and Methods} 41, pp. 3915-3941. 
  }
}
\keyword{datasets}