\name{VAM-package}
\alias{VAM-package}
\docType{package}
\title{
Variance-Adjusted Mahalanobis
}
\description{
Implementation of Variance-adjusted Mahalanobis (VAM), a method for cell-specific gene set scoring of scRNA-seq data.
}
\details{
\tabular{ll}{
Package: \tab VAM\cr
Type: \tab Package\cr
Version: \tab 0.5.0 \cr
Date: \tab 2021\cr
License: \tab GPL-2 \cr
}
}
\author{
H. Robert Frost
}
\references{
    \itemize{
    \item Frost, H. R. (2020). Variance-adjusted Mahalanobis (VAM): a fast and accurate method for cell-specific gene set scoring. biorXiv e-prints. doi: https://doi.org/10.1101/2020.02.18.954321 
    }
}
\note{
This work was supported by the National Institutes of Health grants K01LM012426, R21CA253408, P20GM130454 and P30CA023108.
}
\keyword{ package }

